added textUtils.jsx split stopwords into stopwords and tokenizing
